+++ GameCore/Interfaces/ICollider.cs
@@ -2,11 +2,11 @@
 {
     public interface ICollider
     {
-        float X { get; set; }
-        float Y { get; set; }
-        float Width { get; set; }
-        float Height { get; set; }
-        float HorizontalSpeed { get; set; }
-        float VerticalSpeed { get; set; }
+        int X { get; set; }
+        int Y { get; set; }
+        int Width { get; set; }
+        int Height { get; set; }
+        int HorizontalSpeed { get; set; }
+        int VerticalSpeed { get; set; }
     }
 }
