+++ GameCore/Interfaces/ITextureGetter.cs
@@ -4,8 +4,8 @@ namespace GameCore
 {
     public interface ITexture
     {
-        float TextureOffSetX { get; }
-        float TextureOffSetY { get; }
+        int TextureOffSetX { get; }
+        int TextureOffSetY { get; }
         int TextureWidth { get; }
         int TextureHeight { get; }
         string TextureName { get; }
